Introduce `ElementNaming` service
