+++ src/component/definition/index.ts
@@ -3,3 +3,4 @@ export * from './custom-elements';
 export * from './definition-context';
 export * from './definition-setup';
 export * from './element-def';
+export * from './element-naming';
